Add UserPetition.image attribute
