+++ backend/app/config/config.yml
@@ -169,6 +169,12 @@ knp_gaufrette:
                 bucket_name: '%amazon_s3.bucket%'
                 options:
                     directory: 'post'
+        image_petition_adapter:
+            aws_s3:
+                service_id: 'aws.s3'
+                bucket_name: '%amazon_s3.bucket%'
+                options:
+                    directory: 'petition'
         thumbnail_post_adapter:
             aws_s3:
                 service_id: 'aws.s3'
@@ -198,6 +204,8 @@ knp_gaufrette:
             adapter: image_announcement_adapter
         image_post_fs:
             adapter: image_post_adapter
+        image_petition_fs:
+            adapter: image_petition_adapter
         thumbnail_post_fs:
             adapter: thumbnail_post_adapter
         thumbnail_petition_fs:
@@ -247,6 +255,10 @@ vich_uploader:
         image_post:
             uri_prefix: /post
             upload_destination: image_post_fs
+            namer: vich_uploader.namer_uniqid
+        image_petition:
+            uri_prefix: /petition
+            upload_destination: image_petition_fs
             namer: vich_uploader.namer_uniqid
         thumbnail_post:
             uri_prefix: /thumbnail/post
