+++ backend/src/Civix/CoreBundle/Resources/config/vich_uploader/Entity.UserPetition.yml
@@ -1,4 +1,7 @@
 Civix\CoreBundle\Entity\UserPetition:
+    image.file:
+        mapping: image_petition
+        filename_property: image.name
     facebookThumbnail.file:
         mapping: thumbnail_petition
         filename_property: facebookThumbnail.name
